Shashi, Jaswanth  | Removing symlink only on uninstall but not on upgrade

+++ package/scripts/preuninstall.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-
-rm -rf /var/www/implementer_interface
+if [ $1 -eq 0 ]; then
+    rm -rf /var/www/implementer_interface
+    rm -rf /etc/httpd/conf.d/implementer_interface_ssl.conf
+fi
